requirements.txt
spacy>=1.8.2,<=1.8.99
pytorch>=0.1.12,<=0.1.99
